+++ base_tier_validation/models/tier_definition.py
@@ -58,6 +58,9 @@ class TierDefinition(models.Model):
         "this definition is triggered.",
     )
     has_comment = fields.Boolean(string="Comment", default=False)
+    comment_on_rejection = fields.Boolean(
+        string="Comment only on rejection", default=False
+    )
     approve_sequence = fields.Boolean(
         string="Approve by sequence",
         default=False,